# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WasteWise API",
    "version": "1.0.0",
...
    "contact": {
      "name": "WasteWise Support",
      "email": "support@wastewise.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://wastewise.com/terms"
    }
  },
  "servers": [
    {
      "url": "https://wastewise.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token obtained from /api/auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code"
          },
          "details": {
            "type": "object",
            "description": "Additional error context",
            "additionalProperties": true
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "required": [
          "error",
          "code",
          "details"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "object",
            "properties": {
              "issues": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "AnalysisJob": {
        "type": "object",
        "required": [
          "id",
          "status",
          "job_type",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Job ID"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "job_type": {
            "type": "string",
            "enum": [
              "complete_analysis",
              "invoice_extraction",
              "contract_extraction",
              "regulatory_research"
            ],
            "description": "Type of analysis job"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed"
            ],
            "description": "Current job status"
          },
          "progress_percent": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "description": "Completion percentage"
          },
          "current_step": {
            "type": "string",
            "description": "Current processing step",
            "example": "Extracting invoice data"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "description": "Processing start time"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "description": "Processing completion time"
          },
          "error_message": {
            "type": "string",
            "description": "Error message if job failed"
          },
          "error_code": {
            "type": "string",
            "description": "Error code if job failed"
          },
          "result_data": {
            "type": "object",
            "description": "Analysis results (available when status is completed)",
            "additionalProperties": true
          },
          "ai_usage": {
            "type": "object",
            "description": "AI token usage and cost tracking",
            "properties": {
              "input_tokens": {
                "type": "integer"
              },
              "output_tokens": {
                "type": "integer"
              },
              "total_cost": {
                "type": "number"
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Authentication required",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
...
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Resource not found",
              "code": "NOT_FOUND"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Invalid request data",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Internal server error",
              "code": "INTERNAL_ERROR"
            }
          }
        }
      }
    },
    "parameters": {
...
      "JobIdParam": {
        "name": "id",
        "in": "path",
        "description": "Job ID",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/api/analyze": {
      "post": {
        "summary": "Create analysis job",
        "description": "Creates a background job for long-running waste analysis operations.\n\n**Async Job Pattern**:\n1. Create job with this endpoint → Returns `jobId`\n2. Poll `GET /api/jobs/{jobId}` every 2 seconds\n3. Check `status` field: `pending` → `processing` → `completed` or `failed`\n4. When completed, access results in `result_data` field\n\n**Processing Time**: 30 seconds to 5 minutes depending on job type and data volume\n",
        "tags": [
          "Analysis"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "projectId",
                  "jobType"
                ],
                "properties": {
                  "projectId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Project ID to analyze",
                    "example": "d82e2314-7ccf-404e-a133-0caebb154c7e"
                  },
                  "jobType": {
                    "type": "string",
                    "enum": [
                      "invoice_extraction",
                      "regulatory_research",
                      "complete_analysis",
                      "report_generation"
                    ],
                    "description": "Type of analysis to perform",
                    "example": "complete_analysis"
                  },
                  "inputData": {
                    "type": "object",
                    "description": "Optional job-specific input parameters",
                    "additionalProperties": true,
                    "example": {
                      "includeRegulatory": true
                    }
                  },
                  "priority": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10,
                    "description": "Job priority (1=highest, 10=lowest). Auto-assigned if not provided.",
                    "example": 5
                  }
                }
              },
              "examples": {
                "complete_analysis": {
                  "summary": "Complete analysis with all features",
                  "value": {
                    "projectId": "d82e2314-7ccf-404e-a133-0caebb154c7e",
                    "jobType": "complete_analysis",
                    "inputData": {
                      "includeRegulatory": true,
                      "generateReports": true
                    }
                  }
                },
                "invoice_extraction": {
                  "summary": "Invoice extraction only",
                  "value": {
                    "projectId": "d82e2314-7ccf-404e-a133-0caebb154c7e",
                    "jobType": "invoice_extraction",
                    "priority": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobId": {
                      "type": "string",
                      "format": "uuid",
                      "description": "Job ID for polling status"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "pending"
                      ],
                      "description": "Initial job status"
                    },
                    "message": {
                      "type": "string",
                      "description": "Instructions for polling"
                    }
                  }
                },
                "example": {
                  "jobId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                  "status": "pending",
                  "message": "Analysis job created. Poll /api/jobs/a1b2c3d4-e5f6-7890-abcd-ef1234567890 for status updates."
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Project not found or access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Project not found or access denied"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (10 requests/minute)",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds until rate limit resets"
              },
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Maximum requests per window"
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Remaining requests in current window"
              },
              "X-RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix timestamp when limit resets"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "retryAfter": {
                      "type": "integer"
                    }
                  }
                },
                "example": {
                  "error": "Rate limit exceeded. Please try again later.",
                  "retryAfter": 45
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/jobs/{id}": {
      "get": {
        "summary": "Get job status",
        "description": "Poll this endpoint to check the status of an analysis job.\n\n**Polling Pattern**:\n- Poll every 2 seconds while `status === \"pending\"` or `status === \"processing\"`\n- Stop polling when `status === \"completed\"` or `status === \"failed\"`\n- Use `progress_percent` and `current_step` to show progress to users\n",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/JobIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Job status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisJob"
                },
                "examples": {
                  "pending": {
                    "summary": "Job pending in queue",
                    "value": {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "user_id": "u123",
                      "project_id": "d82e2314-7ccf-404e-a133-0caebb154c7e",
                      "job_type": "complete_analysis",
                      "status": "pending",
                      "progress_percent": 0,
                      "current_step": "Waiting in queue",
                      "created_at": "2025-01-15T10:00:00Z"
                    }
                  },
                  "processing": {
                    "summary": "Job in progress",
                    "value": {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "user_id": "u123",
                      "project_id": "d82e2314-7ccf-404e-a133-0caebb154c7e",
                      "job_type": "complete_analysis",
                      "status": "processing",
                      "progress_percent": 45,
                      "current_step": "Extracting invoice data (3/6 files)",
                      "started_at": "2025-01-15T10:01:00Z",
                      "created_at": "2025-01-15T10:00:00Z"
                    }
                  },
                  "completed": {
                    "summary": "Job completed successfully",
                    "value": {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "user_id": "u123",
                      "project_id": "d82e2314-7ccf-404e-a133-0caebb154c7e",
                      "job_type": "complete_analysis",
                      "status": "completed",
                      "progress_percent": 100,
                      "current_step": "Completed",
                      "started_at": "2025-01-15T10:01:00Z",
                      "completed_at": "2025-01-15T10:04:30Z",
                      "result_data": {
                        "yardsPerDoor": 0.12,
                        "costPerDoor": 18.5,
                        "optimizations": [
                          {
                            "type": "compactor_monitors",
                            "savings": 1200
                          }
                        ]
                      },
                      "ai_usage": {
                        "input_tokens": 15000,
                        "output_tokens": 3000,
                        "total_cost": 0.45
                      },
                      "created_at": "2025-01-15T10:00:00Z"
                    }
                  },
                  "failed": {
                    "summary": "Job failed with error",
                    "value": {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "user_id": "u123",
                      "project_id": "d82e2314-7ccf-404e-a133-0caebb154c7e",
                      "job_type": "complete_analysis",
                      "status": "failed",
                      "progress_percent": 30,
                      "current_step": "Failed during invoice extraction",
                      "error_message": "Invalid invoice format: missing required fields",
                      "error_code": "EXTRACTION_ERROR",
                      "started_at": "2025-01-15T10:01:00Z",
                      "completed_at": "2025-01-15T10:02:00Z",
                      "created_at": "2025-01-15T10:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}